+++ src/renderer/components/file-explorer/FileExplorerModel.tsx
@@ -594,6 +594,12 @@ export class FileExplorerModel extends TComponentModel<FileExplorerState, FileEx
             });
         }
 
+        items.push({
+            label: "Open in New Panel",
+            icon: <NewWindowIcon />,
+            onClick: () => pagesModel.addEmptyPageWithNavPanel(item.filePath),
+        });
+
         if (this.props.enableFileOperations && item.filePath !== this.props.rootPath) {
             items.push(
                 {